+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/packet/PacketIds.java
@@ -25,6 +25,7 @@ public final class PacketIds {
     public static final ResourceLocation RESOURCE_SLOT_AMOUNT_CHANGE = createIdentifier("resource_slot_amount_change");
     public static final ResourceLocation SINGLE_AMOUNT_CHANGE = createIdentifier("detector_amount_change");
     public static final ResourceLocation USE_NETWORK_BOUND_ITEM = createIdentifier("use_network_bound_item");
+    public static final ResourceLocation NETWORK_TRANSMITTER_STATUS = createIdentifier("network_transmitter_status");
 
     private PacketIds() {
     }